+++ contracts/sg-eth-airdrop/src/lib.rs
@@ -1,8 +1,1 @@
-pub mod contract;
-pub mod error;
-pub mod msg;
-pub mod query;
-pub mod reply;
-pub mod state;
-pub use crate::error::ContractError;
-mod claim_airdrop;
+pub mod rekt_airdrop;
