+++ contracts/sg-eth-airdrop/src/rekt_airdrop.rs
@@ -1,0 +1,8 @@
+pub mod claim_airdrop;
+pub mod contract;
+pub mod error;
+pub mod msg;
+pub mod query;
+pub mod reply;
+pub mod state;
+pub mod testing;
