+++ contracts/sg-eth-airdrop/src/rekt_airdrop/error.rs
