+++ contracts/sg-eth-airdrop/src/rekt_airdrop/msg.rs
