+++ contracts/sg-eth-airdrop/src/rekt_airdrop/state.rs
