+++ contracts/sg-eth-airdrop/src/rekt_airdrop/testing.rs
@@ -1,0 +1,6 @@
+// #[cfg(test)]
+// mod constants;
+// #[cfg(test)]
+// mod setup;
+// #[cfg(test)]
+// mod tests;
