--- contracts/sg-eth-airdrop/src/testing.rs
@@ -1,6 +1,0 @@
-#[cfg(test)]
-mod constants;
-#[cfg(test)]
-mod setup;
-#[cfg(test)]
-mod tests;
