+++ test-suite/src/sg_eth_airdrop/tests/test_claim.rs
@@ -7,7 +7,7 @@ use crate::sg_eth_airdrop::setup::setup_signatures::{
 use crate::sg_eth_airdrop::setup::test_msgs::InstantiateParams;
 use async_std::task;
 use cosmwasm_std::{Addr, Attribute, Coin, Uint128};
-use sg_eth_airdrop::msg::{ExecuteMsg, QueryMsg};
+use sg_eth_airdrop::rekt_airdrop::msg::{ExecuteMsg, QueryMsg};
 
 use ethers_core::rand::thread_rng;
 use ethers_signers::{LocalWallet, Signer};
@@ -22,7 +22,7 @@ use crate::sg_eth_airdrop::setup::execute_msg::{
     execute_contract_error_with_msg, execute_contract_with_msg, instantiate_contract,
 };
 
-use sg_eth_airdrop::contract::INSTANTIATION_FEE;
+use sg_eth_airdrop::rekt_airdrop::contract::INSTANTIATION_FEE;
 
 fn query_minter_as_expected(app: &mut StargazeApp, airdrop_contract: Addr, minter_addr: Addr) {
     let query_msg = QueryMsg::GetMinter {};
